refactor(lint): move into tools/lint sub-module
Isolates the lint tool from the framework's go.mod so it can depend
on github.com/a-h/templ for the upcoming url-attr check without
adding templ to every library consumer's go.sum.

No behavior change. go install path becomes
github.com/jackielii/structpages/tools/lint/cmd/structpages-lint@latest.
