+++ tools/lint/analyzer.go
