+++ tools/lint/directive.go
