+++ tools/lint/directive_test.go
