+++ tools/lint/go.mod
@@ -1,0 +1,13 @@
+module github.com/jackielii/structpages/tools/lint
+
+go 1.24.0
+
+require golang.org/x/tools v0.42.0
+
+require (
+	github.com/google/go-cmp v0.7.0 // indirect
+	golang.org/x/mod v0.33.0 // indirect
+	golang.org/x/sync v0.19.0 // indirect
+)
+
+replace github.com/jackielii/structpages => ../..
