+++ tools/lint/idfor.go
