+++ tools/lint/tree.go
