+++ tools/lint/util.go
